requirements.txt
httpx
azul-runner>=4.0.0
Pillow>=10.3

